FIX stupid API for plot_bands
The argument resolution should be provided only for functions for which
there is a need to know the resolution (ie, functions that take as input
data something binned). `plot_bands` and `set_genomic_labels` are not
such functions: the input data should be at the same resolution as the
lengths argument, and the user should just know to provide them at the
same resolution.

closes #46